Recreate WASAPI objs when the default device changes.
The intent here is to gracefully switch the rendering device
when the current rendering device is unplugged.

By using the HRESULT values returned by functions, we can detect
when the current rendering device disappears (e.g. headphones
unplugged).

However, by using only those values we can't detect when the default device
is changed to something else without unplugging the previous device, e.g.
plugging the headphones back in. Hence the use of IMMNotificationClient to
detect changes in the devices.