Change the order of WASAPI, XAudio2 and WinMM.
... to support automatic initialization of the backend better.

WASAPI is a lower level API than WinMM and XAudio2. On Windows,
both XAudio and WinMM internally use WASAPI. WASAPI supposedly
offers lower latency, and is included by default in all Windows versions
after XP - unlike XAudio2, which requires additional DLLs.

However, WASAPI does not offer any conversions when it comes
to buffers, making it a bit demanding to use from SoLoud. WinMM, on
the other hand converts merrily, and is included on all the same
modern platforms as is WASAPI.

Considering all of this, it makes sense to *first* try WASAPI,
and then the others if that fails.